Capture correct tracebacks when using `inline_callbacks`.
When an inlineCallback completes synchronously, such as in tests, twisted will
string-ify the traceback, which means that when the exception is then thrown
into another inlineCallback function, the stack trace is lost[1]. Twisted works
around this by inspect the stack and finding the Failure instance from the
inner call, to get traceback there.

However, this only works if `Failure.throwExceptionIntoGenerator` is used.
This adjusts eliot_friendly_generator_function to use `Failure`, instead of an
exc_info tuple, when used via `inline_callabcks`.


[1] The string-ified traceback can't be passed to `.send`.